refactor(framework): Move EventType and event fn
