+++ framework/py/flwr/common/__init__.py
@@ -54,12 +54,12 @@ from ..compat.common.typing import ServerMessage as ServerMessage
 from ..compat.common.typing import Status as Status
 from ..compat.common.typing import Value as Value
 from ..supercore.date import now as now
+from ..supercore.telemetry import EventType as EventType
+from ..supercore.telemetry import event as event
 from .constant import MessageTypeLegacy as MessageTypeLegacy
 from .grpc import GRPC_MAX_MESSAGE_LENGTH
 from .logger import configure as configure
 from .logger import log as log
-from .telemetry import EventType as EventType
-from .telemetry import event as event
 
 if TYPE_CHECKING:
     from ..app.message import Array as Array
