+++ framework/py/flwr/common/exit/exit.py
@@ -23,11 +23,11 @@ import time
 from logging import ERROR, INFO
 from typing import Any, NoReturn
 
-from flwr.common import EventType, event
 from flwr.supercore.constant import (
     FORCE_EXIT_TIMEOUT_SECONDS,
     TELEMETRY_TIMEOUT_SECONDS,
 )
+from flwr.supercore.telemetry import EventType, event
 from flwr.supercore.version import package_version
 
 from ..logger import log
