+++ framework/py/flwr/supercore/task_process/agent/run_agentapp.py
@@ -42,7 +42,6 @@ from flwr.common.serde import (
     fab_from_proto,
     run_from_proto,
 )
-from flwr.common.telemetry import EventType, event
 from flwr.proto.appio_pb2 import (  # pylint: disable=E0611
     PullTaskInputRequest,
     PullTaskInputResponse,
@@ -54,6 +53,7 @@ from flwr.supercore.superexec.dependency_installer import (
     cleanup_app_runtime_environment,
     install_app_dependencies,
 )
+from flwr.supercore.telemetry import EventType, event
 from flwr.supercore.typing import JSONObject
 from flwr.superlink.grid import GrpcGrid
